song/requirements.txt
aiohttp==3.7.4
async-timeout==3.0.1
chardet==3.0.4
discord.py
multidict==4.5.2
websockets==9.1
configparser==3.7.1
beautifulsoup4==4.7.1
pytz==2018.9
argparse==1.4.0
aiopg
aiodns==2.0.0
aiosqlite==0.9.0
cairosvg
opencv-contrib-python-headless
snscrape
psycopg[binary]>=3.1.0
psycopg_pool